+++ .github/workflows/lint.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Lint with flake8
         run: |
           pip install flake8
-          flake8 ./aws/logs_monitoring/ --count --select=E9,F63,F7,F82 --show-source --statistics
+          flake8 ./aws/logs_monitoring/ --count --select=F4,F7,F8 --show-source --statistics
 
       - name: Check formatting with Black
         run: |
